# llminthehouse-backend/src/main/resources/openapi/llminthehouse-api.yaml
openapi: 3.0.2
info:
  title: llminthehouse API
  description: Routes for managing systems, roles, rights, and user roles
  version: '1.0.0-ALPHA'

servers:
  - url: http://{host}:{port}/{contextPath}
    description: Internal dev server for testing
    variables:
      host:
        default: localhost
      port:
        default: "8100"
      contextPath:
        default:

security:
  - bearerAuth: [ ]

paths:

  /auth/register:
    post:
      operationId: register
      summary: Register new user
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequestWebDTO'
      responses:
        '201':
          description: User registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserWebDTO'
        '400':
          description: Registration failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralErrorWebDTO'

  /auth/login:
    post:
      operationId: login
      summary: User login
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequestWebDTO'
      responses:
        '200':
          description: Successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserWebDTO'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralErrorWebDTO'

  /conversations:
    post:
      operationId: createConversation
      summary: Create Conversation
      tags:
        - Conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConversationWebDTO'
      responses:
        '200':
          description: Successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConversationWebDTO'
        '401':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralErrorWebDTO'

  /conversations/{uuid}:
    post:
      operationId: getConversation
      summary: Get Conversation
      tags:
        - Conversation
      parameters:
        - in: path
          name: uuid
          schema:
            type: string
            format: uuid
          required: true
          description: "uuid"
      responses:
        '200':
          description: Successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConversationWebDTO'
        '401':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralErrorWebDTO'

  /conversations/user/{useruuid}:
    post:
      operationId: getConversationsForUser
      summary: Get Conversation for given user
      tags:
        - Conversation
      parameters:
        - in: path
          name: useruuid
          schema:
            type: string
            format: uuid
          required: true
          description: "uuid"
      responses:
        '200':
          description: Successful
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ConversationListWebDTO'
        '401':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralErrorWebDTO'

  /conversation/{uuid}/message:
    post:
      operationId: addMessage
      summary: Add Message
      tags:
        - Conversation
      parameters:
        - in: path
          name: uuid
          schema:
            type: string
            format: uuid
          required: true
          description: "uuid"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MessageWebDTO'
      responses:
        '200':
          description: Successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageWebDTO'
        '401':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralErrorWebDTO'


components:

  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:

    UserWebDTO:
      type: object
      required:
        - id
        - uuid
        - username
        - createdAt
      properties:
        id:
          type: integer
          format: int64
        uuid:
          type: string
          format: uuid
        username:
          type: string
        createdAt:
          type: string
          format: date-time

    LoginRequestWebDTO:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
        password:
          type: string


    RegisterRequestWebDTO:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
        password:
          type: string

    MessageWebDTO:
      type: object
      required:
        - id
        - content
        - timestamp
        - number
        - llm
      properties:
        id:
          type: integer
          format: int64
        content:
          type: string
        timestamp:
          type: string
          format: date-time
        number:
          type: integer
        llm:
          type: boolean

    ConversationListWebDTO:
      type: object
      required:
        - id
        - uuid
        - title
        - createdAt
      properties:
        id:
          type: integer
          format: int64
        uuid:
          type: string
          format: uuid
        title:
          type: string
        createdAt:
          type: string
          format: date-time

    ConversationWebDTO:
      type: object
      required:
        - id
        - uuid
        - title
        - createdAt
        - user
        - messages
      properties:
        id:
          type: integer
          format: int64
        uuid:
          type: string
          format: uuid
        title:
          type: string
        createdAt:
          type: string
          format: date-time
        user:
          $ref: '#/components/schemas/UserWebDTO'
        messages:
          type: array
          items:
            $ref: '#/components/schemas/MessageWebDTO'

    GeneralErrorWebDTO:
      type: object
      required:
        - message
      properties:
        type:
          type: string
          description: "Type of the error (usually the class name of the exception)"
        message:
          type: string
          description: "Error message"